requirements.txt
PyQt5
Pillow
numpy
pyVulkan